+++ Source/Table/Boundary_Behaviors/return_zero_at_boundary.h
@@ -45,8 +45,8 @@ namespace nemesis
 
 
     // Functionality
-    float lookup(std::size_t index,
-                 float       key) const
+    double lookup(std::size_t index,
+                  double      key) const
     {
       return 0.0;
     }
